requirements.txt
streamlit>=1.30.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
python-dotenv>=1.0.0
groq>=0.4.0
openpyxl>=3.1.2
sqlalchemy>=2.0.0
